fix: change shell from bash to node
