+++ scripts/camelcase.js
@@ -1,5 +1,5 @@
 module.exports = function camelCase(value) {
   return value.replace(/-\w/g, (r, i) => {
-    return value.charAt(i + 1).toUpperCase();
+    return value.charAt(i + 1).toUpperCase()
   })
 }
